+++ src/main/java/com/sparta/goodbite/domain/restaurant/service/RestaurantService.java
@@ -1,12 +1,16 @@
 package com.sparta.goodbite.domain.restaurant.service;
 
+import com.sparta.goodbite.auth.security.EmailUserDetails;
 import com.sparta.goodbite.domain.operatinghour.dto.OperatingHourResponseDto;
 import com.sparta.goodbite.domain.operatinghour.entity.OperatingHour;
 import com.sparta.goodbite.domain.operatinghour.repository.OperatingHourRepository;
+import com.sparta.goodbite.domain.owner.entity.Owner;
 import com.sparta.goodbite.domain.restaurant.dto.RestaurantRequestDto;
 import com.sparta.goodbite.domain.restaurant.dto.RestaurantResponseDto;
 import com.sparta.goodbite.domain.restaurant.entity.Restaurant;
 import com.sparta.goodbite.domain.restaurant.repository.RestaurantRepository;
+import com.sparta.goodbite.exception.restaurant.RestaurantErrorCode;
+import com.sparta.goodbite.exception.restaurant.detail.RestaurantNotAuthorizationException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,9 +24,12 @@ public class RestaurantService {
     private final OperatingHourRepository operatingHourRepository;
 
     @Transactional
-    public void createRestaurant(RestaurantRequestDto restaurantRequestDto) {
+    public void createRestaurant(RestaurantRequestDto restaurantRequestDto,
+        EmailUserDetails userDetails) {
 
-        restaurantRepository.save(restaurantRequestDto.toEntity());
+        Owner owner = (Owner) userDetails.getUser();
+
+        restaurantRepository.save(restaurantRequestDto.toEntity(owner));
     }
 
     @Transactional(readOnly = true)
@@ -43,16 +50,27 @@ public class RestaurantService {
     }
 
     @Transactional
-    public void updateRestaurant(Long restaurantId, RestaurantRequestDto restaurantRequestDto) {
+    public void updateRestaurant(Long restaurantId, RestaurantRequestDto restaurantRequestDto,
+        EmailUserDetails userDetails) {
 
         Restaurant restaurant = restaurantRepository.findByIdOrThrow(restaurantId);
+
+        Owner owner = (Owner) userDetails.getUser();
+
+        checkOwnerByRestaurant(owner, restaurant);
+
         restaurant.update(restaurantRequestDto);
     }
 
     @Transactional
-    public void deleteRestaurant(Long restaurantId) {
+    public void deleteRestaurant(Long restaurantId, EmailUserDetails userDetails) {
 
         Restaurant restaurant = restaurantRepository.findByIdOrThrow(restaurantId);
+
+        Owner owner = (Owner) userDetails.getUser();
+
+        checkOwnerByRestaurant(owner, restaurant);
+
         restaurantRepository.delete(restaurant);
     }
 
@@ -66,5 +84,12 @@ public class RestaurantService {
         return operatingHours.stream()
             .map(OperatingHourResponseDto::from)
             .toList();
+    }
+
+    private void checkOwnerByRestaurant(Owner owner, Restaurant restaurant) {
+        if (!owner.equals(restaurant.getOwner())) {
+            throw new RestaurantNotAuthorizationException(
+                RestaurantErrorCode.RESTAURANT_NOT_AUTHORIZATION);
+        }
     }
 }
